requirements.txt
mattermostdriver==7.3.2
jira==3.5.2
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
asyncio==3.4.3
datetime==5.2
logging==0.4.9.6
schedule==1.2.0 